requirements.txt
telethon
pandas
nltk
pillow
opencv-python
python-dotenv